+++ R/GSODTools-deprecated.R
@@ -1,0 +1,15 @@
+#' Deprecated functions in \pkg{GSODTools}
+#' 
+#' @description 
+#' The functions listed below are deprecated and will be defunct in
+#'   the near future. When possible, alternative functions with similar
+#'   functionality are also mentioned. Help pages for deprecated functions are
+#'   available at \code{help("GSODTools-deprecated")}.
+#'
+#' \itemize{
+#'   \item \code{\link[GSODTools]{gsodReformat-deprecated}}
+#' }   
+#'   
+#' @name GSODTools-deprecated
+#' @keywords internal
+NULL