+++ NEWS.md
@@ -1,3 +1,18 @@
+# GSODTools 1.0.4.9003 (2022-06-08)
+
+#### ✨ features and improvements
+
+#### 🐛 bug fixes
+
+#### 💬 documentation etc
+
+#### 🍬 miscellaneous
+
+  * Eliminates {sp} and {raster} dependencies --> uses {sf} instead (#10)
+  * Disables interactive selection of area of interest in `stationFromExtent()`
+  * Removes obsolete arguments in `gsodReformat()` and deprecates function
+
+
 # GSODTools 1.0.4 (2022-05-18)
 
 #### ✨ features and improvements